pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hybro-hub"
version = "0.1.8"
description = "Hybro Hub — daemon and Gateway client for connecting local and cloud A2A agents on hybro.ai."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Hybro", email = "info@hybro.ai" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27,<1",
    "httpx-sse>=0.4,<1",
    "pydantic>=2.0,<3",
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18",
    "click>=8.0",
    "a2a-sdk>=0.3,<1",
    "psutil>=5.9",
    "a2a-adapter",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[project.scripts]
hybro-hub = "hub.cli:main"

[project.urls]
Homepage = "https://hybro.ai"
Documentation = "https://github.com/hybroai/hybro-hub#readme"
Repository = "https://github.com/hybroai/hybro-hub"

[tool.hatch.build.targets.wheel]
packages = ["hybro_hub", "hub"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
